requirements.txt
# 基础依赖
numpy>=1.21.0
opencv-python>=4.5.0
pandas>=1.3.0
Pillow>=8.0.0

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# YOLOv5依赖
PyYAML>=5.4.1
requests>=2.25.1
scipy>=1.7.0
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0

# CRNN依赖
tensorboard>=2.6.0

# 工具库
pathlib>=1.0.1
typing-extensions>=3.10.0

# 开发工具
pytest>=6.2.5  # 用于测试
black>=21.7b0  # 代码格式化
flake8>=3.9.0  # 代码检查 